+++ CHANGELOG.md
@@ -6,6 +6,7 @@
   - increase test coverage
   - check for Python minimal version 3.6
   - new endpoint `v3/connect/loginFromSession`
+  - new endpoint `v1/card/getClearPin`
 
 
 ## [20191112]